media: imx: vdic: Prepare for mem2mem processing
This patch re-orders VDIC hardware operations to make it possible to
call the ->setup() operation, but not the ->start() operation, at
stream on. This paves the way to allow for the single-shot device_run
processing op to call only ->start() at device_run, and ->stop at
job_done, without needing to also ->setup() and ->disable() for every
job.

Signed-off-by: Steve Longerbeam <slongerbeam@gmail.com>