media: imx: Create imx_media_video_dev_ops
Instead of storing the video device format in struct imx_media_video_dev,
create imx_media_video_dev_ops to retrieve the format. And expand the
operations to include retrieving both output and capture formats, to
support mem2mem devices as well as capture-only devices

Finally, create ops that retrieve a new buffer that is ready for I/O,
and return a completed buffer back to the video device, and an op to flag
errors during streaming. As a result of these new ops, imx_media_buffer
is now only used by imx-media-capture, so make that structure private
to it.

Signed-off-by: Steve Longerbeam <slongerbeam@gmail.com>