media: rcar-vin: Move media_device_register to async completion
The media_device is registered during driver probe, before async
completion, so it is possible for .link_notify to be called before
all sub-devices are bound.

Fix this by moving media_device_register() to rvin_group_notify_complete().
This ensures that all sub-devices are now bound (the rcar-csi2 subdevices
and and video capture devices) before .link_notify can be called.

Fixes: c0cc5aef3170 ("media: rcar-vin: add link notify for Gen3")

Signed-off-by: Steve Longerbeam <slongerbeam@gmail.com>