media: rcar-csi2: Fix field detection
The RCAR Gen3 hardware manual doesn't mention that the "Word Count" field
in the MIPI CSI-2 Frame Start packet is really the frame number. FS packets
are short packets, and short packets do not have a WC but rather a 16-bit
data field, and for Frame synchronization packets (FS/FE) the 16-bit data
field contains a frame number.

Here is a reprinting from the MIPI CSI-2 specification, version 2
(section 9.1.2 Low Level Protocol Short Packet Format):

"Figure 42 and Figure 43 show the Low Level Protocol Short Packet
structures for the D-PHY and C-PHY physical layer options, respectively.
For each option, the Short Packet structure matches the Packet Header of
the corresponding Low Level Protocol Long Packet structure with the
exception that the Packet Header Word Count (WC) field shall be replaced
by the Short Packet Data Field...For Frame Synchronization Data Types
the Short Packet Data Field shall be the frame number..."

Also in section 9.8.1 Frame Synchronization Packets, the CSI-2 spec reads:

"The behavior of the 16-bit frame number shall be as one of the following
- Frame number is always zero – frame number is inoperative.
- Frame number increments by 1 for every FS packet with the same Virtual
  Channel and is periodically reset to one e.g. 1, 2, 1, 2, 1, 2, 1, 2 or
  1, 2, 3, 4, 1, 2, 3, 4"

So from the above, FLD_NUM in the FLD register matches against the
frame number transmitted by the CSI-2 source in the Frame Start Packet,
and goes as "1,2,1,2,1,2" or "1,2,3,4,...,1,2,3,4". If there is a match,
the RCAR CSI-2 receiver declares the field as the even, e.g. the BOTTOM,
field.

NTSC transmits bottom/even field first, and PAL and HD transmit top/odd
field first. So fields with a frame number 1 in the CSI-2 Frame Start
Packet are bottom/even fields for NTSC, and frame number 2 are bottom/even
fields for PAL/HD.

But the above assumes the transmitting CSI-2 sensor is sending frame
numbers in the sequence "1,2,1,2,...". But according to the CSI-2 spec,
sensors are also allowed to send frame numbers that simply increment as
in "1,2,3,4,...". To generalize to catch those cases, set FLD_DET_SEL
to 0x01, which according to the RCAR Gen3 hardware manual means "the
field is detected as the EVEN field when FLD_NUM[0] matches WC[0]".
Then, the even/bottom fields have odd frame numbers for NTSC, and even
frame numbers for PAL (this patch assumes HDMI sources will not
implement .g_std(), so std will be 0 in that case).

Finally, set the FLD register to non-zero value only if trnsmitter is
sending ALTERNATE fields.

Signed-off-by: Steve Longerbeam <slongerbeam@gmail.com>