media: imx: prpencvf: Prepare for mem2mem processing
This patch re-orders IC hardware operations to make it possible to
setup the hardware for streaming, but not actually start streaming,
at stream on. This paves the way to allow for the single-shot device_run
processing op to only start capture at device_run, and stop capture at
job_done, without needing to also setup and un-setup for every job.

Signed-off-by: Steve Longerbeam <slongerbeam@gmail.com>