gpu: ipu-v3: vdic: Fix ipu_vdi_set_field_order()
The description of the TOP_FIELD bits in the imx6 reference manual is
not descriptive enough. The manual states the bits are set as:

0    top field is field 0
1    top field is field 1

It is not clear from that description which field is tagged the top field
by this bit, since in indirect mode, the VDIC receives three fields from
memory over DMA channels 8,9,10: F(n-1), F(n), and F(n+1). So which of
those is tagged the top field?

The answer can be found from a reply to a question on NXP community
site, see [1]. When the bit is 0, F(n-1) is the TOP field (TBT order),
and when the bit is 1, F(n) is the TOP field (BTB order).

By using the field type to determine the TOP field position,
ipu_vdi_set_field_order() is making assumptions about which sequence
of three fields are passed to the VDIC (n-2, n-1, n, or n-1, n, n+1 ??).
Instead, modify the API to ipu_vdi_set_field_order() to take the position
of the TOP field, which is unambiguous:

position 0 = F(n-1) is the TOP field
position 1 = F(n) is the TOP field

[1] https://community.nxp.com/message/472928?commentID=472928#comment-472928

Signed-off-by: Steve Longerbeam <slongerbeam@gmail.com>