gpu: ipu-v3: vdic: Fix ipu_vdi_set_field_order()
The description of the VDI_C_TOP_FIELD bits in the imx6 reference manual
is not clear. The manual states the bits are set as:

0    top field is field 0
1    top field is field 1

It is not clear from that description which field is tagged the top field
by this bit, since in indirect mode, the VDIC receives three fields from
memory over DMA channels 8,9,10: F(n-1), F(n), and F(n+1). So is "field
0" F(n-1) or F(n), and is "field 1" F(n) or F(n+1)?

The answer can be found from a reply to a question on NXP community
site, see [1]. The important piece of info from that posting is the
following:

"
top field is field 0 means there are two odd fields;
top field is field 1 means there are two even fields.
"

In other words, when the bit is 0, there are two odd (TOP) fields, i.e.
TBT order, so F(n-1) is the TOP field. And when the bit is 1, there are
two even (BOTTOM) fields, i.e. BTB order, so F(n) is the TOP field. In
summary, the VDI_C_TOP_FIELD bits are interpreted as:

0  TOP field is F(n-1)
1  TOP field is F(n)

To clear all this up, modify the API to ipu_vdi_set_field_order() to pass
a single enumeration 'enum ipu_vdi_field' that makes it clear which field
is being tagged the TOP field:

IPU_VDI_FIELD_PREV = TOP field is F(n-1)
IPU_VDI_FIELD_CURR = TOP field is F(n)

and program the VDI_C_TOP_FIELD bits correctly according to above.

[1] https://community.nxp.com/message/472928?commentID=472928#comment-472928

Fixes: 2d2ead4530771 ("gpu: ipu-v3: Add Video Deinterlacer unit")
Signed-off-by: Steve Longerbeam <slongerbeam@gmail.com>